Deb: Move my_print_defaults to MariaDB client core package
The my_print_default is required by mytop which is also in the MariaDB
client package, which in turn requires the client core package.
This way it is ensured my_print_default is always available when mytop
is installed.

The my_print_defaults is also requires by the server logrotate files
and mysqld_safe script etc, but this change is fine, since the server
core package always depends on the MariaDB client core package anyway
and they are both installed on server installations.

Also add appropriate Breaks/Replaces.

Commit cherry-picked from upstream fc65417.