Save autopkgtests results as JUnit-compatible XML-report
The mysql-test-run tool supports creating JUnit-compatible XML-reports,
which can be parsed and used by various CI tools. Extend the Salsa-CI
rules to use it so the results are visible on the pipeline Tests-tab.

Also start using eatmydata in the mysql-test-run to make it about 15%
faster, and set parallel=auto for automatic maximum concurrency.