MDEV-10075: Provide index of error causing error in array INSERT
Extended the parser for GET DIAGNOSTICS to use ERROR_INDEX to get
warning/error index. Error information is stored in Sql_condition. So it can
be used to store the index of warning/error too. records variable of
st_copy_info stores the number of records processed when we are 'trying'
to write them to the table, irrespecive if they are eventually inserted or
not. So, its current value can be stored in a thread class variable
(current_insert_index) and assigned to the variable (error_index) of
Sql_condition object when it is created during error/warning. This error_index
variable is returned when ERROR_INDEX is used in GET DIAGNOSTICS.