Salsa-CI: Give mariadb-upgrade more time to complete and ignore Gnitpick
MariaDB 10.2 and 10.3 don't have a mariadb.sys user, so the upgrade needs
to create one and it takes some time. If the database is queried too
quickly after start it might fail with:

  ERROR 1449 (HY000) at line 1: The user specified as a definer
  ('mariadb.sys'@'localhost') does not exist

Also ignore Gnitpick errors as it seems to be a bit buggy with this branch
naming scheme of 'debian/latest'.