Clean away most Lintian overrides and unstable-tests
The Lintian overrides and unstable-tests lists contain
many entries that are several years old. To discover
which ones are still actually needed, I first removed
all of them and added back only the ones that cause
trigger errors in latest version of Lintian or that
fail to pass mariadb-test-run on the architectures
I have access to test on Launchpad.net.

More Lintian overrides and unstable-test lines might
need to be recovered, but it is hard to discover which
unless we first do a upload to Debian experimental to
get full QA exposure.