Remove version suffix from packages, e.g. mariadb-server(-10.6)
Make version suffixed packages (e.g. mariadb-server-10.6) deprecated
as it made maintenance complicated and there was no known use cases or users
of the naming scheme, as multiple different major version MariaDB server
packages could not be co-installed anyway and source or the 'mariadb-server'
and 'mariadb-client' packages is easiest controlled by repositories and
package versioning, not versions in package *names*.

This was already done in upstream branch 10.9 in 375b8f4 (MDEV-27033).