Do not create /var/log/mysql in postinst
Since commit bb847777, error log by default goes to journal/syslog,
so /var/log/mysql remains unused. It hence does not need
to be pre-created. If admins enable file logging, it must be clear
that the chosen directory must exist and be writable to the service
user.

A related hint has been added to the config file and copying
/var/log/mysql has been removed from CI.

Signed-off-by: MichaIng <micha@dietpi.com>