Cover the full-upgrade scenario in CI
User may hit a case where after a full-upgrade, the MariaDB Server is
not working anymore (removed, not started, etc.). See: #1004262. This CI
steps makes sure that after a full-upgrade, the MariaDB Server is still
working and running.

This is done by doing a full-upgrade from Bookworm to Sid (installing
MariaDB packages from artifacts in the full-upgrade process).

Note that doing the full upgrade from Stable (Bullseye as of today) to
Sid is not possible in docker, see #1008202 and the discussion in
https://salsa.debian.org/mariadb-team/mariadb-server/-/merge_requests/18.