Drop obsolete dependency on package lsb-release
The package is now empty and no longer used. This fixes Lintian error:

  E: mariadb-server: depends-on-obsolete-package
  Depends: lsb-base (>= 3.0-10)

The functionality of lsb-base is in the Essential:yes set since
Bullseye. The package itself is now an empty transitional package
(because debootstrap doesn't understand the Provides relationship) which
depends on the new provider of the functionality, sysvinit-utils, which
is also in the Essential:yes set.

More information in
https://lists.debian.org/debian-devel/2023/01/msg00150.html
and https://tracker.debian.org/media/packages/l/lsb/changelog-11.5