Stop skipping tests that are marked fixed upstream
Remove all test overrides for issues that are marked as fixed upstream,
or has a patch in the Debian packaging sources. All build failures after
this are either regressions in previous fixes or new failures.

Also remove all overrides for amd64 as that is the main architecture and
really should have a fully passing main test suite.