Have all compression provider plugins as Recommends for mariadb-server
Previously MariaDB 10.6 and previous versions in Debian included by default
all compression methods. This was also validated by autopkgtests.

Upstream didn't have this, and now MariaDB 10.11 in Debian had the same
provider plugins as upstream. To prevent existing installations from
stopping to work if compression was used, have all compression methods as
Recommends at least for now. Some of them (or all) can be dropped in the
future if one compression method is shown to be superior to others and
having them all in Recommends is later shown to be excessive.