Packages mariadb-plugin-provider-* must depend at least on MariaDB 10.11
The mariadb-plugin-provider-* structure is inherited from upstream, which
started using them in MariaDB 10.9 for both deb and rpm packages. They
are not traditional libraries that we manage with ABI versioning nor do
we track symbols and such, and it is unknown what the long-term use of them
will be by upstream.

However, at minimum these plugins need the MariaDB 10.11 series server in
Debian. If installed with MariaDB 10.6 they will thow an error. Example:

  [ERROR] mariadbd: Can't open shared library 'provider_lz4.so'
  (errno: 8, API version for DAEMON plugin provider_lz4 not supported
  by this version of the server)