Salsa-CI: Adjust Bookworm to Sid upgrades with 1:10.11.2-1 in Bookworm
It is relevant to continue to test MariaDB 10.6 to 10.11 updates
in Debian as there might be users who ran a previous version of
testing or unstable for whom we want to ensure upgrades still work.
Also, testing and maintaining this upgrade path helps future Ubuntu
upgrades as Ubuntu 22.04 did ship with MariaDB 10.6.

However, just one 10.6 upgrade in Bookworm is enough, so remove the
extra one in Sid.

Also document why Buster backport can't be built.