MDEV-31118: Suppress Lintian lsb-base package error
Debian Salsa-CI uses Sid distro Lintian to packages.
From Debian 12 lsb-base package is obsolete and it's
handled correctly on MariaDB LTS 10.11 (which is packaged
in Debian 12 distribution).

MariaDB 10.6-10.10 are targeted to build Debian 10 and 11
this Lintian rule is is false-positive because in these
distros lsb-base is needed.

NOTE TO MERGER: Needed only up-to 10.10