Use native zlib to make crossbuilds work again
The arm64 crossbuild started to fail in MariaDB 10.11.3
as the (libmariadb/external/)zlib/CMakeLists.txt somehow polluted
the main build CMake and made it loose the ZLIB_LIBRARY location
and unable to build with system zlib.

This was visible in the CMake configure stage as messages such as:

    Could NOT find ZLIB (missing: ZLIB_LIBRARY) (found version "1.2.11")
    ...
    SYSTEM_LIBS: ZLIB_LIBRARY-NOTFOUND

The MySQL 8.0.33 adapted to this change with commit
https://github.com/mysql/mysql-server/commit/1f2b9d62#diff-9c0008f51a59d8b1a69304775d8463a8a77a101b3db3e5ad9e77f1f3ba2ae2f6

Our solution is to make the zlib dependency explicitly native.