Salsa-CI: Ensure all sh/bash scripts in debian/* pass Shellcheck
All shell scripts in debian/* are now Shellcheck clean. Start using CI to
enforce that the scripts don't regress.

Also add tiny extra check to ensure debian/rules does not have syntax
errors (but this is not as extensive as Shellcheck).

The new packages add only 3 MB of extra download, so total is now around
50 MB of packages to download. In Salsa-CI all package downloads should
be well cached however.