Fix FTCBFS: Do not pass host CFLAGS to the native build (Closes: #1051222)
MariaDB started failing to cross build from source with a recent dpkg
upload. In that upload, the default CFLAGS for arm64 gained a value that
is incompatible with any other architecture. Passing the arm64 CFLAGS to
any other compiler makes things fail. Now debhelper is asked to recompute
the build flags and as a result it passes the host flags to the native
build.