Bugfix: Rewrite 0300a9157cc to retain original autopkgtest behavior
The command '! dpkg-query -W $plugin' would check if RocksDB exits, and
if not, return zero. If RocksDB did indeed exist, it would fail the test
as a sign that the platform was 64-bit and big endian, and should have
RocksDB but didn't.

This fixed the autopkgtests for 1:10.11.5-1 upload for armhf and i386,
while being a Shellcheck compatible solution.